services/api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
boto3==1.22.3
botocore==1.25.3
click==8.1.2
colorama==0.4.4
fastapi==0.75.2
h11==0.13.0
idna==3.3
jmespath==1.0.0
pydantic==1.9.0
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
gunicorn
